requirements.txt
uvicorn[standard]~=0.22.0
fastapi~=0.91.0
urllib3==2.0.3
openai~=0.27.8
PyYAML~=6.0
requests==2.31.0
aiohttp~=3.8.4
pysondb~=1.6.7
bs4~=0.0.1
beautifulsoup4~=4.12.2
setuptools==68.0.0
dnspython==2.3.0
colorama~=0.4.6
APScheduler~=3.10.1
pytz~=2023.3
pyjwt~=2.7.0
cryptography==41.0.2
starlette~=0.24.0
argparse~=1.4.0
pymongo~=4.4.0
firebase-admin~=6.2.0
sqlalchemy~=2.0.18
stem~=1.8.2
python-multipart